1.add param in HttpHeader's construction function to distinguish http/http2 2.modify param check in encodeInteger

+++ net/http/v2/HPack.cpp
@@ -284,9 +284,9 @@ const byte _HPack::HuffmanCodesLength[] = {
 
 const int _HPack::HUFFMAN_EOS = 256;
 
-const int _HPack::MinHeaderTableSize = 0;
-const int _HPack::MaxHeaderTableSize = 0xffffffffL;
-const int _HPack::DefaultHeaderTableSize = 4096;
-const int _HPack::MaxHeaderListSize = 0xffffffffL;
+const long _HPack::MinHeaderTableSize = 0;
+const long _HPack::MaxHeaderTableSize = 0xffffffffL;
+const long _HPack::DefaultHeaderTableSize = 4096;
+const long _HPack::MaxHeaderListSize = 0xffffffffL;
 
 }
